+++ src/main/kotlin/no/nav/familie/ef/sak/task/PollStatusFraIverksettTask.kt
@@ -1,6 +1,5 @@
 package no.nav.familie.ef.sak.task
 
-import no.nav.familie.ef.sak.repository.domain.Behandling
 import no.nav.familie.ef.sak.service.BehandlingService
 import no.nav.familie.ef.sak.service.steg.StegService
 import no.nav.familie.prosessering.AsyncTaskStep
